Add SwitchSeq sequencer.
Uses the memory space in OC_patterns to sequence. Like a mini sequins.
4 sequences are played in parallel. The note values may be configured with sequins or an upcoming sequence editor. CV can manipulate the sequences in a variety of ways.

There are two channels controlled with a single clock (GATE 1). Gate 2 is a reset to makes the internal sequwncer's step equal to 0.
Each channel has a few different modes, controllable with the encoder.

Modes:
OCT:  you are in this mode when the sequence arrow is solid. The encoder will rotate through sequences and it will play the chosen sequence, with CV controlling the octave (by 1 or 2 octaves).
PICK: In this mode CV will pick which sequence is playing. Sampled after each clock.
QUAN: The internal sequences are ignored and the CV will be quanitzed to a global scale. Right now that's c-dorian and only configurable in code.
RAND: The CV is ignored, and instead a random sequence is chosen at random.